requirements.txt
langchain
openai 
python-dotenv 
todoist-api-python